test(storage): share browser storage fixture
