+++ apps/desktop/src/lib/local-storage.test.ts
@@ -1,25 +1,6 @@
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest'
+import { createStorage } from '@/test-utils/storage'
 import { getLocalStorageStore, resetLocalStorageStores } from './local-storage'
-
-function createStorage(entries: Record<string, string> = {}): Storage {
-  const values = new Map(Object.entries(entries))
-  return {
-    get length() {
-      return values.size
-    },
-    key: (index) => [...values.keys()][index] ?? null,
-    getItem: (key) => values.get(key) ?? null,
-    setItem: (key, value) => {
-      values.set(key, value)
-    },
-    removeItem: (key) => {
-      values.delete(key)
-    },
-    clear: () => {
-      values.clear()
-    },
-  }
-}
 
 beforeEach(() => {
   vi.stubGlobal('window', { localStorage: createStorage({ key: 'local' }) })
