+++ apps/desktop/src/test-utils/storage.ts
@@ -1,0 +1,19 @@
+export function createStorage(entries: Record<string, string> = {}): Storage {
+  const values = new Map(Object.entries(entries))
+  return {
+    get length() {
+      return values.size
+    },
+    key: (index) => [...values.keys()][index] ?? null,
+    getItem: (key) => values.get(key) ?? null,
+    setItem: (key, value) => {
+      values.set(key, value)
+    },
+    removeItem: (key) => {
+      values.delete(key)
+    },
+    clear: () => {
+      values.clear()
+    },
+  }
+}
